requirements.txt
# FastAPI 및 서버
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 템플릿 엔진
jinja2==3.1.3

# 환경 변수 관리
python-dotenv==1.0.0

# HTTP 클라이언트 (API 호출용)
httpx==0.26.0
aiohttp==3.9.1

# 데이터베이스
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
aiomysql==0.2.0
mysqlclient==2.2.1
pymysql==1.1.0

# Redis
redis==5.0.1
aioredis==2.0.1

# Kafka
kafka-python==2.0.2
confluent-kafka==2.3.0

# 데이터 검증
pydantic==2.5.3
pydantic-settings==2.1.0

# 스케줄러
apscheduler==3.10.4

# 데이터 처리
pandas==2.1.4
numpy==1.26.3

# Spark (선택사항 - 로컬 개발시 주석 처리 가능)
# pyspark==3.5.0

# 유틸리티
python-dateutil==2.8.2

# 개발/테스트 도구
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# 코드 품질
black==23.12.1
flake8==7.0.0
